Experiment with Cinemachine and FP/TP Camera Controllers
First person mouse aiming in the sub happens in "steps", very noticeable at a high sensitivity, and (unsuccessful) attempts have been made at tinkering with various parameters to fix it.

Cinemachine has been added to the OnRails scene, but will require further tweaking to get right.

Cinemachine on the Third Person Camera Controller is somewhat difficult to get working right. IIRC for now, it's only needed to handle camera collisions with the environment so it doesn't clip through the floor or other objects.